integrates authentication with UserContext
